feat(metrics): extract log parsing to separate module
- Moved log parsing functions from node_sync_test.py to a new module
  metrics_extractor.py.
- Updated imports in node_sync_test.py to use the new metrics_extractor module.
- Improved code organization and separation of concerns by isolating log
  parsing logic.
